+++ apiserver/facades/agent/storageprovisioner/package_test.go
@@ -5,10 +5,37 @@ package storageprovisioner_test
 
 import (
 	stdtesting "testing"
+	"time"
 
+	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/testing"
+	gc "gopkg.in/check.v1"
 )
 
 func TestAll(t *stdtesting.T) {
 	testing.MgoTestPackage(t)
+}
+
+type storageSetUp interface {
+	setupVolumes(c *gc.C)
+	setupFilesystems(c *gc.C)
+}
+
+const dontWait = time.Duration(0)
+
+type byMachineAndEntity []params.MachineStorageId
+
+func (b byMachineAndEntity) Len() int {
+	return len(b)
+}
+
+func (b byMachineAndEntity) Less(i, j int) bool {
+	if b[i].MachineTag == b[j].MachineTag {
+		return b[i].AttachmentTag < b[j].AttachmentTag
+	}
+	return b[i].MachineTag < b[j].MachineTag
+}
+
+func (b byMachineAndEntity) Swap(i, j int) {
+	b[i], b[j] = b[j], b[i]
 }
